feat(doc): updates to README.md
Add some descriptions of the various pieces of work covered in this repo
and a list of items for things I still need to do.

A section on setup and usage is still required for each of the project
directories. I mainly used the devcontainer to quickly spin up the
necessary docker containers, but I should also clearly include
instructions for how I setup the colima VM with the aarch64
architecture.